feat: ignore any file conaining '@pirates: ignore'
This was the result of a rather intiresting bug: to have pirates ignore them, I put the compilers for the tests in test/fixture/node_modules. That meant that git ignored them. However, if I moved them out of node_modules, they would go through pirates, and I didn't want that. So this fixes that an adds a useful feature!