chore(travis): don't run coverage on travis
Code coverage is badly broken, because istanbul doesn't support
child_processes, and the fork of it that does (nyc) causes an
infinate loop.

Since the creator of nyc was one of the main people involved in the
pirates discussion, I'm hoping that they will eventually switch to
pirates. Then we can figure out a way to fix it.

Maybe we could figure out a way to pre-instrument the files, then a
require hook wouldn't be needed at all.