perf: remove ignore comment regex
Previously, you could include the phrase  anywhere in a file and it would be ignored by pirates. This was generally a bad idea, so now it's been removed. This should also deliver a nice speed boost

BREAKING CHANGE: `/* @pirates: ignore */` now doesn't do anything. Since this was undocumented, you shouldn't have been using it, but if you were, now you're not.